+++ src/views/RegistrationView.js
@@ -1,29 +1,10 @@
 import Media from 'react-media';
-import {
-  Container,
-  Logo,
-  IconContainer,
-  ImageContainer,
-  ImageText,
-  Icon,
-  Form,
-  Input,
-  ButtonSubmit,
-  ButtonLink,
-  LogoText,
-} from './styles';
+import { Container, Logo, ImageContainer, ImageText } from './styles';
 
 import { ReactComponent as LogoTextIcon } from '../icons/logo-text.svg';
-import { ReactComponent as LogoIcon } from '../icons/logo.svg';
-import { ReactComponent as EmailIcon } from '../icons/email.svg';
-import { ReactComponent as PasswordIcon } from '../icons/password.svg';
-import { ReactComponent as NameIcon } from '../icons/name.svg';
+import FormRegistration from '../components/FormRegistration';
 
 export default function LoginView() {
-  const hadleSubmit = e => {
-    e.preventDefault();
-  };
-
   return (
     <Media
       queries={{
@@ -39,38 +20,7 @@ export default function LoginView() {
               <Logo>
                 <LogoTextIcon />
               </Logo>
-              <Form onSubmit={hadleSubmit}>
-                <IconContainer>
-                  <Input type="email" placeholder="E-mail" />
-                  <Icon>
-                    <EmailIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Подтвердите пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="name" placeholder="Ваше имя" />
-                  <Icon>
-                    <NameIcon />
-                  </Icon>
-                </IconContainer>
-
-                <ButtonSubmit type="submit">Регистрация</ButtonSubmit>
-                <ButtonLink type="button">Вход</ButtonLink>
-              </Form>
+              <FormRegistration />
             </Container>
           )}
           {matches.tablet && (
@@ -79,42 +29,7 @@ export default function LoginView() {
                 <img src="/images/registrationTablet.png" alt="Logo" />
                 <ImageText>Finance App</ImageText>
               </ImageContainer>
-              <Form onSubmit={hadleSubmit}>
-                <Logo>
-                  <LogoIcon />
-                  <LogoText>Wallet</LogoText>
-                </Logo>
-                <IconContainer>
-                  <Input type="email" placeholder="E-mail" />
-                  <Icon>
-                    <EmailIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Подтвердите пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer margin="50px">
-                  <Input type="name" placeholder="Ваше имя" />
-                  <Icon>
-                    <NameIcon />
-                  </Icon>
-                </IconContainer>
-
-                <ButtonSubmit type="submit">Вход</ButtonSubmit>
-                <ButtonLink type="button">Регистрация</ButtonLink>
-              </Form>
+              <FormRegistration />
             </Container>
           )}
           {matches.desktop && (
@@ -123,42 +38,7 @@ export default function LoginView() {
                 <img src="/images/registrationDesktop.png" alt="Logo" />
                 <ImageText>Finance App</ImageText>
               </ImageContainer>
-              <Form onSubmit={hadleSubmit}>
-                <Logo>
-                  <LogoIcon />
-                  <LogoText>Wallet</LogoText>
-                </Logo>
-                <IconContainer>
-                  <Input type="email" placeholder="E-mail" />
-                  <Icon>
-                    <EmailIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer>
-                  <Input type="password" placeholder="Подтвердите пароль" />
-                  <Icon>
-                    <PasswordIcon />
-                  </Icon>
-                </IconContainer>
-
-                <IconContainer margin="50px">
-                  <Input type="name" placeholder="Ваше имя" />
-                  <Icon>
-                    <NameIcon />
-                  </Icon>
-                </IconContainer>
-
-                <ButtonSubmit type="submit">Вход</ButtonSubmit>
-                <ButtonLink type="button">Регистрация</ButtonLink>
-              </Form>
+              <FormRegistration />
             </Container>
           )}
         </>
